Finally working stat parsing
Finalized the PyMongo script that I wrote to parse and send data to the Mongo DB. We finished the parsing for basic yardage last semester, but had issues with authentication to the DB and the syntax for finding and updating documents. Now have basic yardage for offensive players in the test DB that can be seen on the website, going to work on getting more comprehensive stats so we can begin our data analysis.