Created new parse debugger
Uses python to go through and compare each play to various cases, printing to a file for ones that don't match a pattern. Using this I will add more and more patterns to catch edge cases, then convert it to adding correct stats for players instead of writing to a file once finished.